Update before move
I'm committing the working lever (remember to undo the freeze positions), switch needs a simple change to be pushable instead of grabbable and the working button with animation. Also every gear in the room has been animated. Honestly don't know where any of the other animations that the rest of the group promised is but John has done the lightning for the tesla coil I just need to add it in once he's done re-texturing it.